pyproject.toml
[tool.poetry]
name = "data-analysis"
version = "0.1.0"
description = ""
authors = ["jh1950 <rlawnsgl191@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
aiosqlite = "^0.20.0"
jinja2 = "^3.1.4"
pandas = "^2.2.2"
openpyxl = "^3.1.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
